--- frontend/app/lib/features/bible/widgets/app.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "../../../.."
-		},
-		{
-			"path": "../../../../../../backend"
-		}
-	],
-	"settings": {}
-}
