+++ frontend/app/lib/widgets/bible_reader/reader_top_bar.dart
